pyproject.toml
[project]
name = "iceaggr"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "pandas>=2.3.3",
    "polars>=1.34.0",
    "pyarrow>=21.0.0",
    "pytorch-lightning>=2.0.0",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "wandb>=0.15.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
